sum_greater_than
This challenge took me 30-45 mins but made me feel good after completing it . It was easy when I did it. I do had made a little imperfection by putting that final answer print statement in the fpr loop giving the result I don't want. But when I put that in the function and not in the loop I got only the final answer and nothing extra.